// generated/dto/onboard-meta.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Auth API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface OnboardMeta
 */
export interface OnboardMeta {
    /**
     * 
     * @type {boolean}
     * @memberof OnboardMeta
     */
    'is_onboarded'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OnboardMeta
     */
    'profile_complete'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OnboardMeta
     */
    'workspace_create'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OnboardMeta
     */
    'workspace_invite'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OnboardMeta
     */
    'workspace_join'?: boolean;
}